Correct the clustering of nodes
The way that the nodes (places and transitions) in the petri net
were clustered with the help of SNAKES was wrong. Now, the cluster is not overwritten anymore, but correctly added in the intended way.
This resolves a bug where the deletion of specific nodes resulted in an KeyError in the snakes library.